+++ Source/SwiftLintFramework/Rules/RuleConfigurations/RegexConfiguration.swift
@@ -1,7 +1,7 @@
 import Foundation
 import SourceKittenFramework
 
-public struct RegexConfiguration: RuleConfiguration, Equatable, CacheDescriptionProvider {
+public struct RegexConfiguration: RuleConfiguration, Hashable, CacheDescriptionProvider {
     public let identifier: String
     public var name: String?
     public var message = "Regex matched."
@@ -74,5 +74,9 @@ public struct RegexConfiguration: RuleConfiguration, Equatable, CacheDescription
         if let severityString = configurationDict["severity"] as? String {
             try severityConfiguration.apply(configuration: severityString)
         }
+    }
+
+    public func hash(into hasher: inout Hasher) {
+        hasher.combine(identifier)
     }
 }